+++ inc/functions/setup.php
@@ -63,11 +63,16 @@ if ( ! function_exists( 'storefront_setup' ) ) :
 		) );
 
 		/*
-		 * Switch default core markup for search form, comment form, and comments
+		 * Switch default core markup for search form, comment form, comments, galleries, captions and widgets
 		 * to output valid HTML5.
 		 */
 		add_theme_support( 'html5', array(
-			'search-form', 'comment-form', 'comment-list', 'gallery', 'caption',
+			'search-form',
+			'comment-form',
+			'comment-list',
+			'gallery',
+			'caption',
+			'widgets',
 		) );
 
 		// Setup the WordPress core custom background feature.