+++ changelog.md
@@ -9,6 +9,8 @@
 * **Dev** - Replaced `previous_post_link()` and `next_post_link()` with `the_post_navigation()`.
 * **Dev** - HTML5 widget support.
 * **Dev** - Fixed typo in classname `storefront-feautred-products` is now `storefront-featured-products`.
+* **Dev** - Updated node-bourbon
+* **Dev** - Replaced instances of `box-sizing` mixin with standard css.
 
 ### *2015.03.09* - 1.4.0
 * **New** - Added a 'Handheld' menu location.