+++ changelog.md
@@ -1,6 +1,6 @@
 # Storefront Changelog
 
-### *2015.03.10* - 1.4.1
+### *2015.03.24* - 1.4.1
 * **Fix** - Remove unnecessary `!important` declaration on star rating color.
 * **Fix** - Star rating display in IE11.
 * **Fix** - Site header margin when using shop page as homepage.
@@ -8,13 +8,14 @@
 * **Tweak** - Layout selector graphics.
 * **Tweak** - Accessibility improvements in post meta.
 * **Tweak** - Products widget styling.
-* **Tweak** - Widget headings are now `h2`s.
+* **Tweak** - Widget headings are now `h3`s.
 * **Tweak** - Skip links are now a function (`storefront_skip_links()`) hooked into `storefront_header`.
 * **Tweak** - Header widget region markup only displays when widgets are assigned.
+* **Tweak** - `:focus` styles.
 * **Dev** - Replaced `paginate_links()` with `the_posts_pagination()`.
 * **Dev** - Replaced `previous_post_link()` and `next_post_link()` with `the_post_navigation()`.
 * **Dev** - HTML5 widget support.
-* **Dev** - Fixed typo in classname `storefront-feautred-products` is now `storefront-featured-products`.
+* **Dev** - Fixed typo in classname: `storefront-feautred-products` is now `storefront-featured-products`.
 * **Dev** - Updated node-bourbon
 * **Dev** - Replaced instances of `box-sizing` mixin with standard css.